Implement decimal quantity and stock handling for invoices
- Modified `invoice_item.quantity` to `decimal(15,2)`.
- Updated `Invoice` model to cast pivot `quantity` to `decimal:2`.
- Updated `Item` model `stock` field to `decimal:2` (already partially done, ensured consistency).
- Created `InvoiceStockService` to handle all stock adjustments (deduction, restoration, individual adjustments) for invoice items, supporting decimal quantities.
- Refactored Filament `CreateInvoice` and `EditInvoice` pages to use `InvoiceStockService` for stock management.
- Updated Filament `InvoiceResource` form to allow decimal input for item quantities and correctly calculate totals with decimals.
- Ensured Filament bulk delete/restore actions also use `InvoiceStockService`.
- Rewrote `InvoiceFeatureTest` to focus on testing `InvoiceStockService` with decimal quantities.
- Deleted obsolete `InvoiceObserverTest.php`.
- Installed necessary PHP extensions (sqlite3, intl) for the testing environment.

+++ app/Traits/InvoiceCalculationTrait.php
@@ -32,7 +32,7 @@ trait InvoiceCalculationTrait
         foreach ($itemsData as $item) {
             if (!empty($item['item_id']) && isset($item['price']) && isset($item['quantity'])) {
                 $price = self::parseCurrencyValue($item['price']);
-                $quantity = (int)($item['quantity'] ?? 1);
+                $quantity = (float)($item['quantity'] ?? 1.0); // Changed to float
                 $subtotal += $price * $quantity;
             }
         }
